+++ old-ui/app/app.js
@@ -35,8 +35,8 @@ const Import = require('./accounts/import')
 const ForgetDeviceScreen = require('./components/connect-hardware/forget-screen')
 import ConnectHardwareForm from './components/connect-hardware/index'
 const InfoScreen = require('./info')
-const PaymentsScreen = require('../../ui/app/Rif/Pages/paymentsPage')
-const DomainsScreen = require('../../ui/app/Rif/Pages/domainsPage')
+const PaymentsScreen = require('../../ui/app/rif/pages/paymentsPage')
+const DomainsScreen = require('../../ui/app/rif/pages/domainsPage')
 const AppBar = require('./components/app-bar/app-bar.component')
 const Loading = require('./components/loading')
 const BuyView = require('./components/buy-button-subview')
