+++ ui/app/rif/pages/domainsPage.js
