+++ ui/app/rif/pages/paymentsPage.js
